full game loop

+++ Assets/Scripts/Games/Minigame.cs
@@ -433,7 +433,7 @@ namespace HeavenStudio.Games
             return null;
         }
 
-        public void ScoreMiss(double weight = 1f)
+        public void ScoreMiss(float weight = 1f)
         {
             double beat = Conductor.instance?.songPositionInBeatsAsDouble ?? -1;
             GameManager.instance.ScoreInputAccuracy(beat, 0, true, NgLateTime(), weight, false);
